requirements.txt
# Discord
discord.py[voice]>=2.3.2
PyNaCl>=1.5.0

# Audio extraction
yt-dlp>=2024.1.0

# Async HTTP
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
humanize>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
